Branch recipenew

+++ src/main/java/duke/list/recipelist/RecipeIngredientList.java
@@ -30,7 +30,7 @@ public class RecipeIngredientList {
         return arrList;
     }
 
-    public void addRecipeIngredient(String ingredientName, int quantity, String weight) throws ParseException {
+    public void addRecipeIngredient(String ingredientName, double quantity, String weight) throws ParseException {
         recipeIngredientList.add(new RecipeIngredient(ingredientName, quantity, weight));
         int index = recipeIngredientList.size();
         if (index == 1) {
